@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,6 +28,8 @@ dependencies:
   google_sign_in:
   firebase_storage: ^3.1.5
   cloud_firestore: ^0.13.4+2
+  image_picker: ^0.6.4
+  cached_network_image: ^2.0.0
 
 dev_dependencies:
   flutter_test:
