@@ -1,19 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:fluttertripapp/User/model/user.dart';
 
 class Place {
+
   String id;
   String name;
   String description;
   String urlImage;
   int likes;
-  User userOwner;
+  bool liked;
+  //User userOwner;
 
-  Place(
-      {Key key,
-      @required this.name,
-      @required this.description,
-      @required this.urlImage,
-      this.likes,
-      @required this.userOwner});
+  Place({
+    Key key,
+    @required this.name,
+    @required this.description,
+    @required this.urlImage,
+    @required this.likes,
+    @required this.liked,
+    this.id
+    //@required this.userOwner
+  });
 }
