@@ -1,8 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
 import 'package:fluttertripapp/Place/model/place.dart';
 import 'package:fluttertripapp/User/model/user.dart';
+import 'package:fluttertripapp/User/ui/widgets/profile_place.dart';
 
 class CloudFirestoreAPI {
   final String USERS = "users";
@@ -19,19 +19,92 @@ class CloudFirestoreAPI {
       'email': user.email,
       'photoUrl': user.photoUrl,
       'myPlaces': user.myPlaces,
-      'myfavoritePlaces': user.myFavoritePlaces,
+      'myFavoritePlaces': user.myFavoritePlaces,
       'lastSignIn': DateTime.now()
     }, merge: true);
   }
 
   Future<void> updatePlaceData(Place place) async {
     CollectionReference refPlaces = _db.collection(PLACES);
+
     await _auth.currentUser().then((FirebaseUser user) {
       refPlaces.add({
         'name': place.name,
         'description': place.description,
         'likes': place.likes,
-        'userOwner': "${USERS}/${user.uid}"
+        'urlImage': place.urlImage,
+        'liked': place.liked,
+        'userOwner': _db.document("${USERS}/${user.uid}"), //reference
+      }).then((DocumentReference dr) {
+        dr.get().then((DocumentSnapshot snapshot) {
+          //ID Place REFERENCIA ARRAY
+          DocumentReference refUsers = _db.collection(USERS).document(user.uid);
+          refUsers.updateData({
+            'myPlaces': FieldValue.arrayUnion(
+                [_db.document("${PLACES}/${snapshot.documentID}")])
+          });
+        });
+      });
+    });
+  }
+
+  List<ProfilePlace> buildMyPlaces(List<DocumentSnapshot> placesListSnapshot) {
+    List<ProfilePlace> profilePlaces = List<ProfilePlace>();
+    placesListSnapshot.forEach((p) {
+      profilePlaces.add(ProfilePlace(
+        Place(
+            name: p.data['name'],
+            description: p.data['description'],
+            urlImage: p.data['urlImage'],
+            likes: p.data['likes'],
+            liked: p.data['liked']),
+      ));
+    });
+
+    return profilePlaces;
+  }
+
+  List<Place> buildPlaces(
+      List<DocumentSnapshot> placesListSnapshot, User user) {
+    List<Place> places = List<Place>();
+
+    placesListSnapshot.forEach((p) {
+      Place place = Place(
+          id: p.documentID,
+          name: p.data["name"],
+          description: p.data["description"],
+          urlImage: p.data["urlImage"],
+          likes: p.data["likes"],
+          liked: p.data["liked"]);
+      List usersLikedRefs = p.data["usersLiked"];
+      place.liked = false;
+      usersLikedRefs?.forEach((drUL) {
+        if (user.uid == drUL.documentID) {
+          place.liked = true;
+        }else{
+          place.liked = false;
+        }
+      });
+      places.add(place);
+    });
+    return places;
+  }
+
+  Future likePlace(Place place, String uid) async {
+    await _db
+        .collection(PLACES)
+        .document(place.id)
+        .get()
+        .then((DocumentSnapshot ds) {
+      int likes = ds.data["likes"];
+      bool liked = ds.data["liked"];
+
+      _db.collection(PLACES).document(place.id).updateData({
+        'likes': place.liked ? likes + 1 : likes - 1,
+        'liked': !liked,
+        'usersLiked': place.liked
+            ? FieldValue.arrayUnion([_db.document("${USERS}/${uid}")])
+            : FieldValue.arrayRemove([_db.document("${USERS}/${uid}")])
       });
     });
   }
