@@ -1,54 +1,62 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:fluttertripapp/Place/model/place.dart';
 import 'package:fluttertripapp/User/bloc/bloc_user.dart';
 import 'package:fluttertripapp/User/model/user.dart';
 import 'package:fluttertripapp/User/ui/widgets/button_bar.dart';
-import 'package:fluttertripapp/User/ui/widgets/user_info.dart';
+import 'package:fluttertripapp/User/ui/widgets/my_user_info.dart';
 import 'package:generic_bloc_provider/generic_bloc_provider.dart';
 
 class ProfileHeader extends StatelessWidget {
-  BlocUser blocUser;
   User user;
+
+  ProfileHeader(@required this.user);
 
   @override
   Widget build(BuildContext context) {
-    blocUser = BlocProvider.of(context);
-    return StreamBuilder(
-      stream: blocUser.streamFirebase,
-      builder: (BuildContext context, AsyncSnapshot snapshot) {
-        switch (snapshot.connectionState) {
-          case ConnectionState.waiting:
-            return CircularProgressIndicator();
-          case ConnectionState.none:
-            return CircularProgressIndicator();
-          case ConnectionState.active:
-            return showProfileData(snapshot);
-          case ConnectionState.done:
-            return showProfileData(snapshot);
-        }
-        return null;
-      },
+    final title = Text(
+      "Profile",
+      style: TextStyle(
+          fontFamily: "Lato",
+          color: Colors.white,
+          fontWeight: FontWeight.bold,
+          fontSize: 30.0),
+    );
+
+    return Container(
+      margin: EdgeInsets.only(left: 20.0, right: 20.0, top: 50.0),
+      child: Column(
+        children: <Widget>[
+          Row(
+            children: <Widget>[title],
+          ),
+          MyUserInfo(user),
+          ButtonsBar()
+        ],
+      ),
     );
   }
 
-  Widget showProfileData(AsyncSnapshot snapShot) {
-    if (!snapShot.hasData || snapShot.hasError) {
-      print("No legeado");
+  Widget showProfileData(AsyncSnapshot snapshot) {
+    if (!snapshot.hasData || snapshot.hasError) {
+      print("No logeado");
       return Container(
         margin: EdgeInsets.only(left: 20.0, right: 20.0, top: 50.0),
         child: Column(
           children: <Widget>[
             CircularProgressIndicator(),
-            Text("No se pudo carcar la información. Has login"),
+            Text("No se pudo cargar la información. Haz login")
           ],
         ),
       );
     } else {
       print("Logeado");
-      print(snapShot.data);
+      print(snapshot.data);
       user = User(
-          name: snapShot.data.displayName,
-          email: snapShot.data.email,
-          photoUrl: snapShot.data.photoUrl);
+          name: snapshot.data.displayName,
+          email: snapshot.data.email,
+          photoUrl: snapshot.data.photoUrl);
+
       final title = Text(
         'Profile',
         style: TextStyle(
@@ -65,7 +73,7 @@ class ProfileHeader extends StatelessWidget {
             Row(
               children: <Widget>[title],
             ),
-            UserInfo(user),
+            MyUserInfo(user),
             ButtonsBar()
           ],
         ),
