@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:fluttertripapp/Place/ui/screens/add_place_screen.dart';
 import 'package:fluttertripapp/User/bloc/bloc_user.dart';
 import 'package:generic_bloc_provider/generic_bloc_provider.dart';
+import 'package:image_picker/image_picker.dart';
 import 'circle_button.dart';
 
 class ButtonsBar extends StatelessWidget {
@@ -21,11 +22,14 @@ class ButtonsBar extends StatelessWidget {
                 Color.fromRGBO(255, 255, 255, 0.6), () => {}),
             CircleButton(
                 false, Icons.add, 40.0, Color.fromRGBO(255, 255, 255, 1), () {
-              File image = null;
-              Navigator.push(context,
-                  MaterialPageRoute(
-                      builder: (BuildContext context) =>
-                          AddPlaceScreen(image: image)));
+              ImagePicker.pickImage(source: ImageSource.camera)
+                  .then((File image) {
+                Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                        builder: (BuildContext context) =>
+                            AddPlaceScreen(image: image)));
+              }).catchError((onError) => print(onError));
             }),
 //            Cierra sesion
             CircleButton(
@@ -33,7 +37,7 @@ class ButtonsBar extends StatelessWidget {
                 Icons.exit_to_app,
                 20.0,
                 Color.fromRGBO(255, 255, 255, 0.6),
-                () => {blocUser.signOuth()}),
+                () => {blocUser.signOut()}),
 //            CircleButton(true, Icons.person, 20.0, Color.fromRGBO(255, 255, 255, 0.6))
           ],
         ));
