@@ -1,32 +1,35 @@
 import 'package:flutter/material.dart';
+import 'package:fluttertripapp/User/bloc/bloc_user.dart';
+import 'package:fluttertripapp/User/model/user.dart';
+import 'package:generic_bloc_provider/generic_bloc_provider.dart';
 import 'profile_place.dart';
 import '../../../Place/model/place.dart';
 
 class ProfilePlacesList extends StatelessWidget {
-  Place place = Place(
-      name: "Knuckles Mountains Range",
-      description: "Hiking. Water fall hunting. Natural bath",
-      urlImage:
-          "https://images.unsplash.com/photo-1519681393784-d120267933ba?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1500&q=80",
-      likes: 3);
-  Place place2 = Place(
-      name: "Mountains",
-      description:
-          "Hiking. Water fall hunting. Natural bath’, 'Scenery & Photography",
-      urlImage:
-          "https://images.unsplash.com/photo-1524654458049-e36be0721fa2?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1500&q=80",
-      likes: 10);
+  BlocUser blocUser;
+  final User user;
+  ProfilePlacesList(@required this.user);
 
   @override
   Widget build(BuildContext context) {
+    blocUser = BlocProvider.of<BlocUser>(context);
     return Container(
       margin: EdgeInsets.only(top: 10.0, left: 20.0, right: 20.0, bottom: 10.0),
-      child: Column(
-        children: <Widget>[
-          ProfilePlace(place),
-          ProfilePlace(place2),
-        ],
-      ),
+      child: StreamBuilder(
+          stream: blocUser.myPlacesListSream(user.uid),
+          builder: (context, AsyncSnapshot snapshot) {
+            switch (snapshot.connectionState) {
+              case ConnectionState.none:
+              case ConnectionState.waiting:
+                return CircularProgressIndicator();
+              case ConnectionState.active:
+              case ConnectionState.done:
+              default:
+                return Column(
+                  children: blocUser.buildMyPlaces(snapshot.data.documents),
+                );
+            }
+          }),
     );
   }
 }
